* commit 'e44c16ab73ac66d9a60898c92cc28ae0dab4d626':
  DM-12660 Handle long values for primary keys in herd entities and model-api objects
  DM-12660 Handle long values for primary keys in herd entities and model-api objects

+++ herd-code/herd-dao/src/main/java/org/finra/herd/dao/BusinessObjectDefinitionDao.java
@@ -47,7 +47,7 @@ public interface BusinessObjectDefinitionDao extends BaseJpaDao
      *
      * @return the list of all business object definition entities.
      */
-    public List<BusinessObjectDefinitionEntity> getAllBusinessObjectDefinitionsByIds(List<Integer> ids);
+    public List<BusinessObjectDefinitionEntity> getAllBusinessObjectDefinitionsByIds(List<Long> ids);
 
     /**
      * Gets a business object definition by key.
