+++ herd-code/herd-model/src/main/java/org/finra/herd/model/jpa/NotificationActionEntity.java
@@ -46,18 +46,18 @@ public class NotificationActionEntity extends AuditableEntity
     @Column(name = TABLE_NAME + "_id")
     @GeneratedValue(generator = TABLE_NAME + "_seq")
     @SequenceGenerator(name = TABLE_NAME + "_seq", sequenceName = TABLE_NAME + "_seq", allocationSize = 1)
-    private Integer id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "ntfcn_rgstn_id", referencedColumnName = "ntfcn_rgstn_id", nullable = false)
     private NotificationRegistrationEntity notificationRegistration;
 
-    public Integer getId()
+    public Long getId()
     {
         return id;
     }
 
-    public void setId(Integer id)
+    public void setId(Long id)
     {
         this.id = id;
     }
